Move CreateNext & EditNext to Create & Edit
