+++ src/pages/invoices/edit/Edit.tsx
@@ -8,15 +8,12 @@
  * @license https://www.elastic.co/licensing/elastic-license
  */
 
-import { useClientResolver } from 'common/hooks/clients/useClientResolver';
-import { useCurrentCompany } from 'common/hooks/useCurrentCompany';
+import { InvoiceStatus } from 'common/enums/invoice-status';
 import { useTitle } from 'common/hooks/useTitle';
 import { Client } from 'common/interfaces/client';
-import { Invoice } from 'common/interfaces/invoice';
 import { InvoiceItemType } from 'common/interfaces/invoice-item';
 import { ValidationBag } from 'common/interfaces/validation-bag';
-import { useBlankInvoiceQuery } from 'common/queries/invoices';
-import { blankInvitation } from 'common/stores/slices/invoices/constants/blank-invitation';
+import { useInvoiceQuery } from 'common/queries/invoices';
 import { Page } from 'components/Breadcrumbs';
 import { Default } from 'components/layouts/Default';
 import { Spinner } from 'components/Spinner';
@@ -25,7 +22,8 @@ import { useAtom } from 'jotai';
 import { cloneDeep } from 'lodash';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useSearchParams } from 'react-router-dom';
+import { generatePath, useParams, useSearchParams } from 'react-router-dom';
+import { v4 } from 'uuid';
 import { invoiceAtom, invoiceSumAtom } from '../common/atoms';
 import { ClientSelector } from '../common/components/ClientSelector';
 import { InvoiceDetails } from '../common/components/InvoiceDetails';
@@ -35,110 +33,54 @@ import { InvoiceTotals } from '../common/components/InvoiceTotals';
 import { ProductsTable } from '../common/components/ProductsTable';
 import { useProductColumns } from '../common/hooks/useProductColumns';
 import { useTaskColumns } from '../common/hooks/useTaskColumns';
-import { useHandleCreate } from './hooks/useHandleCreate';
-import { useInvoiceUtilities } from './hooks/useInvoiceUtilities';
+import { useInvoiceUtilities } from '../create/hooks/useInvoiceUtilities';
+import { ActionsNext } from './components/ActionsNext';
+import { useHandleSave } from './hooks/useInvoiceSave';
 
-export type ChangeHandler = <T extends keyof Invoice>(
-  property: T,
-  value: Invoice[typeof property]
-) => void;
-
-export function CreateNext() {
+export function Edit() {
   const { t } = useTranslation();
-  const { documentTitle } = useTitle('new_invoice');
-
-  const [invoice, setInvoice] = useAtom(invoiceAtom);
-
-  const { data } = useBlankInvoiceQuery({
-    enabled: typeof invoice === 'undefined',
-  });
-
-  const clientResolver = useClientResolver();
-  const company = useCurrentCompany();
-
-  const productColumns = useProductColumns();
-  const taskColumns = useTaskColumns();
-
-  const [invoiceSum] = useAtom(invoiceSumAtom);
-
+  const { id } = useParams();
   const [searchParams] = useSearchParams();
-  const [errors, setErrors] = useState<ValidationBag>();
-  const [client, setClient] = useState<Client | undefined>();
 
   const pages: Page[] = [
     { name: t('invoices'), href: '/invoices' },
     {
-      name: t('new_invoice'),
-      href: '/invoices/create',
+      name: t('edit_invoice'),
+      href: generatePath('/invoices/:id/edit', { id }),
     },
   ];
 
+  const productColumns = useProductColumns();
+  const taskColumns = useTaskColumns();
+
+  const { documentTitle } = useTitle('edit_invoice');
+  const { data } = useInvoiceQuery({ id });
+
+  const [invoice, setInvoice] = useAtom(invoiceAtom);
+  const [invoiceSum] = useAtom(invoiceSumAtom);
+
+  const [client] = useState<Client | undefined>();
+  const [errors, setErrors] = useState<ValidationBag>();
+
   const {
     handleChange,
-    calculateInvoiceSum,
     handleInvitationChange,
+    calculateInvoiceSum,
     handleLineItemChange,
     handleLineItemPropertyChange,
     handleCreateLineItem,
     handleDeleteLineItem,
   } = useInvoiceUtilities({ client });
 
-  const save = useHandleCreate(setErrors);
-
   useEffect(() => {
-    console.log(typeof invoice);
-    console.log(data);
-
-    if (typeof data !== 'undefined' && typeof invoice === 'undefined') {
+    if (data) {
       const _invoice = cloneDeep(data);
 
-      if (company && company.enabled_tax_rates > 0) {
-        _invoice.tax_name1 = company.settings.tax_name1;
-        _invoice.tax_rate1 = company.settings.tax_rate1;
-      }
-
-      if (company && company.enabled_tax_rates > 1) {
-        _invoice.tax_name2 = company.settings.tax_name2;
-        _invoice.tax_rate2 = company.settings.tax_rate2;
-      }
-
-      if (company && company.enabled_tax_rates > 2) {
-        _invoice.tax_name3 = company.settings.tax_name3;
-        _invoice.tax_rate3 = company.settings.tax_rate3;
-      }
-
-      if (typeof _invoice.line_items === 'string') {
-        _invoice.line_items = [];
-      }
+      _invoice.line_items.map((lineItem) => (lineItem._id = v4()));
 
       setInvoice(_invoice);
     }
-
-    return () => {
-      setInvoice(undefined);
-    };
   }, [data]);
-
-  useEffect(() => {
-    invoice &&
-      invoice.client_id.length > 1 &&
-      clientResolver.find(invoice.client_id).then((client) => {
-        setClient(client);
-
-        const invitations: Record<string, unknown>[] = [];
-
-        client.contacts.map((contact) => {
-          if (contact.send_email) {
-            const invitation = cloneDeep(blankInvitation);
-
-            invitation.client_contact_id = contact.id;
-            invitations.push(invitation);
-          }
-        });
-
-        handleChange('invitations', invitations);
-      });
-  }, [invoice?.client_id]);
 
   useEffect(() => {
     // The InvoiceSum takes exact same reference to the `invoice` object
@@ -148,13 +90,19 @@ export function CreateNext() {
     invoice && calculateInvoiceSum();
   }, [invoice]);
 
+  const save = useHandleSave(setErrors);
+
   return (
     <Default
       title={documentTitle}
       breadcrumbs={pages}
       onBackClick="/invoices"
-      onSaveClick={() => save(invoice as Invoice)}
-      disableSaveButton={invoice?.client_id.length === 0}
+      onSaveClick={() => invoice && save(invoice)}
+      disableSaveButton={
+        invoice &&
+        (invoice.status_id === InvoiceStatus.Cancelled || invoice.is_deleted)
+      }
+      navigationTopRight={invoice && <ActionsNext />}
     >
       <div className="grid grid-cols-12 gap-4">
         <ClientSelector
@@ -162,8 +110,8 @@ export function CreateNext() {
           onChange={(id) => handleChange('client_id', id)}
           onClearButtonClick={() => handleChange('client_id', '')}
           onContactCheckboxChange={handleInvitationChange}
-          readonly={searchParams.get('table') === 'tasks'}
           errorMessage={errors?.errors.client_id}
+          readonly
         />
 
         <InvoiceDetails invoice={invoice} handleChange={handleChange} />
@@ -238,7 +186,7 @@ export function CreateNext() {
       <div className="my-4">
         {invoice && (
           <InvoicePreview
-            for="create"
+            for="invoice"
             resource={invoice}
             entity="invoice"
             relationType="client_id"
